added abstract class Scheduler and SecondScheduler.java as well as some AvailabilityMatrix methods
