+++ src/main/java/com/example/demo/models/AvailabilityMatrix.java
@@ -10,6 +10,7 @@ import lombok.Getter;
 import java.time.Duration;
 import java.time.LocalTime;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Random;
 
@@ -368,6 +369,34 @@ public class AvailabilityMatrix {
             sb.append("|\n");
         }
         return sb.toString();
+    }
+
+    public List<Candidate> getAllAvailableCandidates(CourseSession courseSession) {
+        int numberOfSlots = courseSession.getDuration() / DURATION_PER_SLOT;
+        List<Candidate> possibleCandidates = new ArrayList<>();
+        for (int i = 0; i < DAYS_IN_WEEK; i++) {
+            for (int j = 0; j < SLOTS_PER_DAY; j++) {
+                if (matrix[i][j] == CourseSession.PREFERRED) {
+                    if (isSlotsAvailable(i, j, numberOfSlots, false)) {
+                        possibleCandidates.add(new Candidate(this, i, j, courseSession.getDuration(), true));
+                    }
+                }
+            }
+        }
+        return possibleCandidates;
+    }
+
+    public List<Candidate> getAllAvailableCandidatesOfDay(CourseSession courseSession, int day) {
+        int numberOfSlots = courseSession.getDuration() / DURATION_PER_SLOT;
+        List<Candidate> possibleCandidates = new ArrayList<>();
+        for (int j = 0; j < SLOTS_PER_DAY; j++) {
+            if (matrix[day][j] == CourseSession.PREFERRED) {
+                if (isSlotsAvailable(day, j, numberOfSlots, false)) {
+                    possibleCandidates.add(new Candidate(this, day, j, courseSession.getDuration(), true));
+                }
+            }
+        }
+        return possibleCandidates;
     }
 }
