+++ src/main/java/com/example/demo/models/TimeTable.java
@@ -3,7 +3,7 @@ package com.example.demo.models;
 import com.example.demo.models.enums.Semester;
 import com.example.demo.models.base.TimestampedEntity;
 import com.example.demo.models.enums.Status;
-import com.example.demo.scheduling.Scheduler;
+import com.example.demo.scheduling.FirstScheduler;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -31,7 +31,7 @@ public class TimeTable extends TimestampedEntity implements Persistable<Long>, S
     private List<CourseSession> courseSessions;
 
     @Transient
-    private Scheduler scheduler;
+    private FirstScheduler firstScheduler;
 
     public TimeTable(){
         this.roomTables = new ArrayList<>();
